Warning batzuk kenduta (int -> NSInteger). Mac OS 10.5 SDK-rekin konpilatu eta exekutagarri unibertsala sortu konpilatzeko orduan (PPC, x86 eta x86_64)

+++ trunk/GrosoSub/Controllers/ASSHeadersController.m
@@ -117,7 +117,7 @@
 }
 
 #pragma mark NSTableDataSource informal protocol
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex
 {
 	ASSHeader *headers = [[self document] headers];
 	
@@ -130,7 +130,7 @@
 	}
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
 	return [[self document] countHeaders];
 }
